Changed endpoint to new version
We've released a new version of SearchForImages. The only difference here is that the response will no longer include values for RefinementOptions. We've done this to decrease the size of the response payload to speed things up. Check the documentation for the possible RefinementOptionsSet values. 

V1 of the SearchForImages endpoint is still valid and supported at this time.